Switch to a better assembly procedure.
The Gaalop binary zip archive is now assembled in the dedicated maven submodule
distribution. This submodule can depend on all other submodules to ensure all
artifacts are available when we want to assemble. This procedure is recommended
here: https://maven.apache.org/plugins/maven-assembly-plugin/faq.html#module-binaries

Additionally the assembled archive now automatically contains the correct
versions of any jars we depend on using maven's dependency mechanism. Previously
all jars in the folder additionalJars were copied to the target assembly during
the build. This requires that any required jar file is manually placed in this
folder in the right version. It is much more natural to let maven handle this.
It already knows all dependencies required for compilation and their right
version.